Fixed meal name extraction from the HTML element.
StartEdu has introduced a breaking change in the HTML structure
of the meal.

The service now add ingredients information the the `.meal-preview` node.
```html
<li>
  NUGGETSY Z KURCZAKA, DIP ŚMIETANOWO-CZOSNKOWY, ZIEMNIAKI OPIEKANE, COLESLAW
  <div class="no-print">
    <i class="small"
      >Skład: 380 gr, biała kapusta, marchew, majonez, jogurt naturalny</i
    >
  </div>
  <div class="allergen-info" preview-status="true">
    <div class="mark-1" title="Zawiera gluten"></div>
    <div class="mark-3" title="Zawiera jajka"></div>
    <div class="mark-7" title="Zawiera mleko / nabiał"></div>
    <img
      class="meal-preview"
      src="https://bsacharski.github.io/startedu-meals/images/IMG_5121_thumb.JPG"
    />
  </div>
</li>
```

As such, the `innerText` property of the node would also return the text with
ingredients information.

The logic was fixed, to pull the information from the first childNode of the
HTML element, which contains only pure text-information about the meal-name.